refactor: rename "user" directory to "users"
