+++ app/journals/journals.validator.ts
@@ -1,5 +1,5 @@
 import { JournalsRepository } from "./journals-repository.type";
-import { UsersRepository } from "../user/users-repository";
+import { UsersRepository } from "../users/users-repository";
 import {
    CreateJournalRequest,
    DeleteJournalRequest,
@@ -25,7 +25,7 @@ export const JournalsValidator = (
 ) : JournalsValidator => ({
    validateGetJournalRequest : async (request : GetJournalRequest) : Promise<ValidationError | null> => {
       if (!request.user)
-         return ValidationError('Unauthorized', 'Unauthorized user.');
+         return ValidationError('Unauthorized', 'Unauthorized users.');
       if (!request.id)
          return ValidationError('BadRequest', 'Journal ID required.');
       if (!(await journalsRepository.exists(request.id)))
@@ -37,7 +37,7 @@ export const JournalsValidator = (
 
    validateGetJournalsRequest : async (request : GetJournalsRequest): Promise<ValidationError | null> => {
       if (!request.user)
-         return ValidationError('Unauthorized', 'Unauthorized user.');
+         return ValidationError('Unauthorized', 'Unauthorized users.');
       if (!(await usersRepository.isAdmin(request.user.username)) && (request.filter.author || request.filter.authorRegex))
          return ValidationError('Forbidden', 'Insufficient permissions.');
       if (request.filter) {
@@ -100,7 +100,7 @@ export const JournalsValidator = (
 
    validateCreateJournalRequest : async (request : CreateJournalRequest): Promise<ValidationError | null> => {
       if (!request.user)
-         return ValidationError('Unauthorized', 'Unauthorized user.');
+         return ValidationError('Unauthorized', 'Unauthorized users.');
       if (!request.name)
          return ValidationError('BadRequest', 'Journal name required.');
       if (!request.author)
@@ -110,7 +110,7 @@ export const JournalsValidator = (
 
    validateUpdateJournalRequest : async (request : UpdateJournalRequest): Promise<ValidationError | null> => {
       if (!request.user)
-         return ValidationError('Unauthorized', 'Unauthorized user.');
+         return ValidationError('Unauthorized', 'Unauthorized users.');
       if (!request.id)
          return ValidationError('BadRequest', 'Journal ID required.');
       if (!request.name)
@@ -124,7 +124,7 @@ export const JournalsValidator = (
 
    validateDeleteJournalRequest : async (request : DeleteJournalRequest): Promise<ValidationError | null> => {
       if (!request.user)
-         return ValidationError('Unauthorized', 'Unauthorized user.');
+         return ValidationError('Unauthorized', 'Unauthorized users.');
       if (!request.id)
          return ValidationError('BadRequest', 'Journal ID required.');
       if (!(await journalsRepository.ownsJournal(request.user.username, request.id)) && !(await usersRepository.isAdmin(request.user.username)))
@@ -136,7 +136,7 @@ export const JournalsValidator = (
 
    validateDeleteJournalsRequest : async (request : DeleteJournalsRequest): Promise<ValidationError | null> => {
       if (!request.user)
-         return ValidationError('Unauthorized', 'Unauthorized user.');
+         return ValidationError('Unauthorized', 'Unauthorized users.');
       if (!(await usersRepository.isAdmin(request.user.username)) && (request.filter.author || request.filter.authorRegex))
          return ValidationError('Forbidden', 'Insufficient permissions.');
       if (request.filter.name && request.filter.nameRegex)
